requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
dotenv==0.9.9
humanfriendly==10.0
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
markdown==3.7
pdfplumber==0.11.5
pandas==2.2.3
extract_msg
python-docx==1.1.2
scikit-learn==1.6.1
tabulate==0.9.0
openpyxl==3.1.5
